fix: move test sync server internally
